Main class for Machine Learning Requirements
class ZoneController:
Main class for machine learning based zone categorization. We use clustering to assign clusters to given locations with known positive cases. We then compute severity of outbreak in a cluster and accordingly assign the threat level. To compute this, we use patient data such as positive cases along with cluster data such as population count in a cluster.